Address feedback about tests
